+++ src/otoole/preprocess/__init__.py
@@ -1,0 +1,5 @@
+from .excel_to_osemosys import generate_csv_from_excel
+from .create_datapackage import main as create_datapackage
+from .narrow_to_datafile import main as create_datafile
+
+__all__ = ['generate_csv_from_excel', 'create_datapackage', 'create_datafile']