+++ CHANGELOG.rst
@@ -6,3 +6,7 @@ Version 0.1
 ===========
 
 - Add CPLEX to csv or CBC solution file conversion script
+- Create CSV files in a folder from an excel workbook
+- Create a Tabular Data Package from a folder of CSVs
+- Create an OSeMOSYS datafile from a Tabular Data Package
+- Adds a command line interface to access these tools